requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy[asyncio]
asyncpg
alembic
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pydantic-settings
email-validator
